moving heuristics
