+++ br.usp.each.saeg.jaguar.core/src/main/java/br/usp/each/saeg/jaguar/core/heuristic/TarantulaHeuristic.java
@@ -1,6 +1,5 @@
-package br.usp.each.saeg.jaguar.core.heuristic.impl;
+package br.usp.each.saeg.jaguar.core.heuristic;
 
-import br.usp.each.saeg.jaguar.core.heuristic.Heuristic;
 
 public class TarantulaHeuristic implements Heuristic {
