+++ br.usp.each.saeg.jaguar.core/src/test/java/br/usp/each/saeg/jaguar/AllTests.java
@@ -2,7 +2,7 @@ package br.usp.each.saeg.jaguar;
 
 import org.junit.runner.RunWith;
 
-import br.usp.each.saeg.jaguar.core.heuristic.impl.OchiaiHeuristic;
+import br.usp.each.saeg.jaguar.core.heuristic.OchiaiHeuristic;
 import br.usp.each.saeg.jaguar.core.runner.JaguarRunnerHeuristic;
 import br.usp.each.saeg.jaguar.core.runner.JaguarSuite;
