+++ br.usp.each.saeg.jaguar.core/src/test/java/br/usp/each/saeg/jaguar/builder/CodeForestXMLBuilderTest.java
@@ -12,7 +12,7 @@ import br.usp.each.saeg.jaguar.codeforest.model.Method;
 import br.usp.each.saeg.jaguar.codeforest.model.Package;
 import br.usp.each.saeg.jaguar.codeforest.model.Requirement;
 import br.usp.each.saeg.jaguar.core.builder.CodeForestXmlBuilder;
-import br.usp.each.saeg.jaguar.core.heuristic.impl.TarantulaHeuristic;
+import br.usp.each.saeg.jaguar.core.heuristic.TarantulaHeuristic;
 import br.usp.each.saeg.jaguar.core.model.core.TestRequirement;
 
 public class CodeForestXMLBuilderTest {
