+++ br.usp.each.saeg.jaguar.core/src/test/java/br/usp/each/saeg/jaguar/heuristic/JaccardTest.java
@@ -4,7 +4,7 @@ import org.junit.Assert;
 import org.junit.Test;
 
 import br.usp.each.saeg.jaguar.core.heuristic.Heuristic;
-import br.usp.each.saeg.jaguar.core.heuristic.impl.JaccardHeuristic;
+import br.usp.each.saeg.jaguar.core.heuristic.JaccardHeuristic;
 
 public class JaccardTest {
